[ENTESB-5323] [KARAF-4523] JMXSecurityMBean bulk canInvoke should be robust even if bulkQuery contains duplicate operations
This fix is based on the original patch by Grzegorz Grzybek, but is modified to just skip the duplicate
method only and continue the entire process so that it can return the complete RBAC info. The original
patch terminated the rest of the process when a duplicate was found.

(cherry picked from commit 1fff0131967ac11b0a481f503fbc952f777fc30d)